No change in behavior: Just pulling workspace details down inside component
